@@ -2,52 +2,54 @@
 	"name": "developer-wiki",
 	"version": "0.0.1",
 	"private": true,
+	"type": "module",
 	"scripts": {
 		"dev": "vite dev",
 		"build": "vite build",
 		"preview": "vite preview",
 		"check": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json",
 		"check:watch": "svelte-kit sync && svelte-check --tsconfig ./tsconfig.json --watch",
-		"lint": "prettier --plugin-search-dir . --check . && eslint .",
-		"format": "prettier --plugin-search-dir . --write . && eslint . --fix"
+		"lint": "prettier --check . && eslint .",
+		"format": "prettier --write . && eslint . --fix"
+	},
+	"dependencies": {
+		"jsdom": "^25.0.1"
 	},
 	"devDependencies": {
 		"@fluent/bundle": "^0.18.0",
 		"@fluent/langneg": "^0.7.0",
 		"@hbsnow/rehype-sectionize": "^1.0.7",
-		"@iconify-json/fa6-brands": "^1.1.19",
-		"@iconify-json/fa6-solid": "^1.1.21",
-		"@iconify-json/tabler": "^1.1.113",
+		"@iconify-json/fa6-brands": "^1.2.1",
+		"@iconify-json/fa6-solid": "^1.2.1",
+		"@iconify-json/tabler": "^1.2.7",
 		"@jsdevtools/rehype-toc": "^3.0.2",
-		"@nubolab-ffwd/svelte-fluent": "^0.8.0",
-		"@sveltejs/adapter-static": "^3.0.1",
-		"@sveltejs/kit": "^2.5.10",
-		"@sveltejs/vite-plugin-svelte": "^3.1.1",
-		"@types/node": "^20.14.0",
-		"@types/prismjs": "^1.26.4",
-		"@typescript-eslint/eslint-plugin": "^5.62.0",
-		"@typescript-eslint/parser": "^5.62.0",
-		"@unocss/extractor-svelte": "^0.57.7",
-		"eslint": "^8.57.0",
-		"eslint-config-prettier": "^8.10.0",
-		"eslint-plugin-import": "^2.29.1",
-		"eslint-plugin-svelte": "^2.39.0",
-		"jsdom": "^24.1.0",
-		"mdsvex": "^0.11.0",
-		"prettier": "^2.8.8",
-		"prettier-plugin-svelte": "^2.10.1",
+		"@nubolab-ffwd/svelte-fluent": "^1.0.0",
+		"@sveltejs/adapter-static": "^3.0.6",
+		"@sveltejs/kit": "^2.7.3",
+		"@sveltejs/vite-plugin-svelte": "^4.0.0",
+		"@types/eslint": "^9.6.1",
+		"@types/node": "^20.17.3",
+		"@types/prismjs": "^1.26.5",
+		"@unocss/extractor-svelte": "^0.63.6",
+		"eslint": "^9.13.0",
+		"eslint-config-prettier": "^9.1.0",
+		"eslint-plugin-svelte": "^2.46.0",
+		"globals": "^15.11.0",
+		"mdsvex": "^0.12.3",
+		"prettier": "^3.3.3",
+		"prettier-plugin-svelte": "^3.2.7",
 		"prismjs": "^1.29.0",
-		"quilt-bulma": "^0.0.23",
-		"rehype-autolink-headings": "^6.1.1",
-		"rehype-rewrite": "^3.0.6",
+		"quilt-bulma": "^0.0.24",
+		"rehype-autolink-headings": "^7.1.0",
+		"rehype-rewrite": "^4.0.2",
 		"rehype-slug": "^6.0.0",
-		"svelte": "^4.2.17",
-		"svelte-check": "^3.8.0",
-		"tslib": "^2.6.2",
-		"typescript": "^5.4.5",
-		"unocss": "^0.57.7",
-		"vite": "^5.2.12",
-		"yaml": "^2.4.3"
-	},
-	"type": "module"
+		"svelte": "^5.1.9",
+		"svelte-check": "^4.0.5",
+		"tslib": "^2.8.0",
+		"typescript": "^5.6.3",
+		"typescript-eslint": "^8.12.2",
+		"unocss": "^0.63.6",
+		"vite": "^5.4.10",
+		"yaml": "^2.6.0"
+	}
 }
