@@ -1,17 +1,18 @@
 <script lang="ts">
-	export let title: string;
-	export let style: string;
-	export let href: string;
-	export let icon: string | undefined = undefined;
+	interface Props {
+		title: string;
+		style: string;
+		href: string;
+		icon?: string;
+	}
+
+	let { title, style, href, icon }: Props = $props();
 </script>
 
 <a {title} {href} class="button is-rounded {style}" target="_blank">
 	{#if icon}
 		<span class="icon-text">
-			<div class={icon} />
-			<slot />
+			<div class={icon}></div>
 		</span>
-	{:else}
-		<slot />
 	{/if}
 </a>
