@@ -3,16 +3,20 @@
 
 	import type { Category } from "./types";
 
-	import { currentLocale } from "$l10n";
+	import current from "$lib/current.svelte";
 
-	export let categories: Category[] = [];
-	export let url: string;
+	interface Props {
+		categories: Category[];
+		url: string;
+	}
+
+	let { categories = [], url }: Props = $props();
 </script>
 
 <div class="column is-narrow">
 	<div class="is-hidden-tablet">
 		<label class="button is-primary" for="toggle-sidebar">
-			<span class="icon"><div class="i-fa6-solid-ellipsis" /></span>
+			<span class="icon"><div class="i-fa6-solid-ellipsis"></div></span>
 			<span>
 				<Localized id="menu" />
 			</span>
@@ -30,24 +34,26 @@
 			<ul class="menu-list">
 				{#each categories as category}
 					<li class:is-hidden={category.draft}>
-						<Localized id={category.slug} let:text let:attrs>
-							{text}
-							<ul>
-								{#each category.pages as page}
-									<li class:is-hidden={page.draft}>
-										<a
-											href={`/${$currentLocale}/${category.slug}/${page.slug}`}
-											class:is-active={url === `${category.slug}/${page.slug}`}
-										>
-											{attrs[page.slug]}
-										</a>
-										<!-- Needed to tell SvelteKit to generate the wiki page with no locale set (with the default locale) -->
-										<a class="is-hidden" href={`/${category.slug}/${page.slug}`}>
-											{attrs[page.slug]}
-										</a>
-									</li>
-								{/each}
-							</ul>
+						<Localized id={category.slug}>
+							{#snippet children({ text, attrs })}
+								{text}
+								<ul>
+									{#each category.pages as page}
+										<li class:is-hidden={page.draft}>
+											<a
+												href={`/${current.locale}/${category.slug}/${page.slug}`}
+												class:is-active={url === `${category.slug}/${page.slug}`}
+											>
+												{attrs[page.slug]}
+											</a>
+											<!-- Needed to tell SvelteKit to generate the wiki page with no locale set (with the default locale) -->
+											<a class="is-hidden" href={`/${category.slug}/${page.slug}`}>
+												{attrs[page.slug]}
+											</a>
+										</li>
+									{/each}
+								</ul>
+							{/snippet}
 						</Localized>
 					</li>
 				{/each}
