@@ -1,22 +1,22 @@
 <script lang="ts">
-	import { localize } from "@nubolab-ffwd/svelte-fluent";
+	import { getFluentContext } from "@nubolab-ffwd/svelte-fluent";
 
-	import { isRtl } from "$l10n";
+	import { localesDir } from "$l10n";
 
-	export let data;
+	const { localize } = getFluentContext();
+
+	let { data } = $props();
 </script>
 
 <svelte:head>
-	<title>{$localize("application-title")}</title>
+	<title>{localize("application-title")}</title>
 </svelte:head>
 
-{#if data.content}
-	{#if !data.translated}
-		<!-- Display the landing page in the correct direction for the fallback locale -->
-		<div dir={isRtl(data.fallbackLocale) ? "rtl" : "ltr"}>
-			<svelte:component this={data.content} />
-		</div>
-	{:else}
-		<svelte:component this={data.content} />
-	{/if}
+{#if !data.translated}
+	<!-- Display the landing page in the correct direction for the fallback locale -->
+	<div dir={localesDir[data.fallbackLocale]}>
+		<data.content />
+	</div>
+{:else}
+	<data.content />
 {/if}
