@@ -1,12 +1,12 @@
 <script lang="ts">
 	import NavbarLangDropdown from "./NavbarLangDropdown.svelte";
 
-	import { currentLocale } from "$l10n";
+	import current from "$lib/current.svelte";
 </script>
 
 <nav class="navbar is-dark" aria-label="main navigation" data-controller="navbar">
 	<div class="navbar-brand no-style">
-		<a class="navbar-item navbar-logo" href={`/${$currentLocale}`}>
+		<a class="navbar-item navbar-logo" href={`/${current.locale}`}>
 			<img src="https://quiltmc.org/assets/img/logo.svg" alt="Quilt Logo" height="28" width="28" />
 
 			<span class="ml-2">DevWiki</span>
@@ -18,11 +18,11 @@
 			data-tooltip="Toggle Low Contrast"
 		>
 			<span id="theme-normal-contrast">
-				<div class="i-fa6-solid-circle-half-stroke i-xl" />
+				<div class="i-fa6-solid-circle-half-stroke i-xl"></div>
 			</span>
 
 			<span id="theme-low-contrast">
-				<div class="i-tabler-circle-half i-xl" />
+				<div class="i-tabler-circle-half i-xl"></div>
 			</span>
 		</label>
 
@@ -30,14 +30,15 @@
 			class="navbar-item px-2 has-tooltip-arrow has-tooltip-bottom"
 			href="https://quiltmc.org/"
 			data-tooltip="Website Home"
+			aria-label="Website Home"
 		>
-			<div class="i-fa6-solid-house i-xl" />
+			<div class="i-fa6-solid-house i-xl"></div>
 		</a>
 
 		<label class="navbar-burger" for="toggle-navbar" aria-label="menu">
-			<span aria-hidden="true" />
-			<span aria-hidden="true" />
-			<span aria-hidden="true" />
+			<span aria-hidden="true"></span>
+			<span aria-hidden="true"></span>
+			<span aria-hidden="true"></span>
 		</label>
 	</div>
 
@@ -52,7 +53,7 @@
 				class="navbar-item has-icon"
 				target="_blank"
 			>
-				<span class="icon"><div class="i-fa6-brands-github i-xl" /></span>
+				<span class="icon"><div class="i-fa6-brands-github i-xl"></div></span>
 				<span class="is-hidden-desktop">GitHub</span>
 			</a>
 		</div>
