@@ -1,23 +1,17 @@
 import { negotiateLanguages } from "@fluent/langneg";
 import { error } from "@sveltejs/kit";
-import { get } from "svelte/store";
 
-import type { ComponentType, SvelteComponent } from "svelte";
+import type { Component } from "svelte";
 
-import { type Locale, defaultLocale, currentLocales, currentLocale } from "$l10n";
+import { type Locale, defaultLocale } from "$l10n";
+import current from "$lib/current.svelte";
 
 export async function load({ params, parent }): Promise<
 	{
-		content?: ComponentType<SvelteComponent>;
+		content?: Component;
 		title: string;
 		draft: boolean;
-	} & (
-		| { translated: true }
-		| {
-				translated: false;
-				fallbackLocale: Locale;
-		  }
-	)
+	} & ({ translated: true } | { translated: false; fallbackLocale: Locale })
 > {
 	// Get the data loaded in the parent layouts
 	const parent_data = await parent();
@@ -38,34 +32,34 @@ export async function load({ params, parent }): Promise<
 		error(404);
 	}
 
-	if (article.availableLocales.includes(get(currentLocale))) {
+	if (article.availableLocales.includes(current.locale)) {
 		// Load the content of the article for the current locale if available
 		const content = await import(
-			`../../../../../wiki/${category.slug}/${article.slug}/${get(currentLocale)}.md`
-		);
+			`../../../../../wiki/${category.slug}/${article.slug}/${current.locale}.md`
+		).then((c) => c.default);
 
 		return {
 			title: `${category.slug}.${article.slug}`,
 			draft: article.draft,
-			content: content.default,
+			content,
 			translated: true
 		};
 	} else if (article.availableLocales.length) {
 		// Load the content of the article for one of the user's prefered locales
 		// (taking the assumptions that if the article exists in a locale, it should exist in english)
-		const fallbackLocale = negotiateLanguages(get(currentLocales), article.availableLocales, {
+		const fallbackLocale = negotiateLanguages(current.locales, article.availableLocales, {
 			defaultLocale,
 			strategy: "lookup"
 		})[0] as Locale;
 
 		const content = await import(
 			`../../../../../wiki/${category.slug}/${article.slug}/${fallbackLocale}.md`
-		);
+		).then((c) => c.default);
 
 		return {
 			title: `${category.slug}.${article.slug}`,
 			draft: article.draft,
-			content: content.default,
+			content,
 			translated: false,
 			fallbackLocale
 		};
