@@ -1,4 +1,4 @@
-import svelteFluent from "@nubolab-ffwd/svelte-fluent/rollup-plugin";
+import svelteFluent from "@nubolab-ffwd/svelte-fluent/vite";
 import { sveltekit } from "@sveltejs/kit/vite";
 import extractorSvelte from "@unocss/extractor-svelte";
 import UnoCSS from "unocss/vite";
